requirements.txt
# NeuronOS Development Requirements

# Core dependencies
PyYAML>=6.0
libvirt-python>=9.0.0

# GUI (optional)
PySide6>=6.5.0

# File monitoring (optional)
inotify-simple>=1.3.0

# Development tools
pytest>=7.0
pytest-cov>=4.0
black>=23.0
mypy>=1.0
ruff>=0.1.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0
